Set type attribute of the clear button to button
A button in a form gets the type attribute set to submit by default, if the type is not specified.
Submitting a form with a clearable datepicker component by pressing enter in a field currently triggers the clearing of the datepicker component, instead of submitting the form.
This can be fixed by setting the type attribute of the clear button to "button".