fix: remove unneeded code to prevent date change with `showTimeSelectOnly`
Before this PR, the below function calls:
    `index.tsx:handleChange -> date_utils:parseDate -> date-fns:parse`
were passing `new Date()` to date-fns parse function as the 'reference date' parameter.
This meant that when the datepicker had the `showTimeSelectOnly` prop and the date format was just a time (e.g. H:mm) the parse result's year/month/day would be today instead of the previously selected date, which we want.

To prevent this problem, the block of code being removed in the present commit would take the current selected date and reset the time to the newly input/parsed time, so that as a result, the year/month/day would not change.

Instead, this PR fixes the problem at its root, by passing `this.props.selected` instead of `new Date()` to the date-fns `parse` function as the reference date, such that the parse result of just a time string will have the same year/month/day as `this.props.selected`. This is the desired behavior, and so this block of code is no longer needed.

This is already tested in test "when update the datepicker input text while props.showTimeSelectOnly is set and dateFormat has only time related format".