fix(types): specify portal return type for render method
This fix solves the warning printed by rollup during build of types,
which is preventing the generation of the portal.d.ts file.

Specifycally, the TS error is the following:
TS2742: The inferred type of render cannot be named without a reference
to react-dom/ node_modules/@types/ react. This is likely not portable.
A type annotation is necessary.

Fixes #5129